Refactor constant file import
