+++ apps/console/src/features/remote-repository-configuration/constants/index.ts
@@ -15,3 +15,5 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+
+export * from "./remote-fetch-constants";
